refactor(store): extract UI, selection, and navigation slices
Three more domain slices, all pure-state with no workspace-mutation
helpers required:

- ui-slice: panel/dialog open flags (left/right/search/command/canvas-
  settings/add-element/highlighter/views/create-view-dialog), canvas
  mode toggles (minimap, snapToGrid, multiSelect, presentation), and
  the pending-delete confirmation pair.

- selection-slice: selectedElementIds / selectedRelationshipId /
  selectedGroupId plus their setters; selecting any object closes the
  Highlighter panel so it doesn't stack with the Inspector.

- navigation-slice: activeViewKey, viewHistory, focusElementId,
  pendingZoomConfirm, createViewDefaults, and the navigation actions
  (setActiveView, drillInto, zoomInto, confirmZoomCreate, etc.).

workspace.ts: 990 → 842 LOC. 969/969 vitest passing.
